alx_travel_app/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14.0
drf-yasg>=1.21.7
django-filter>=23.2
